Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -3,7 +3,7 @@
 int main() {
     int n;
     scanf("%d",&n);
-    int len[n],width[n];
+    float len[n],width[n];
     for(int i=0;i<n;i++){
         scanf("%d %d",&len[i],&width[i]);
     }
